+++ root/programs/CS/Samples/AsyncSvc_sample/TestAsyncSvc_Sample/TestAsyncSvc_Sample.csproj
@@ -85,12 +85,6 @@
       <Install>true</Install>
     </BootstrapperPackage>
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\AsyncSvc_sample\AsyncSvc_sample.csproj">
-      <Project>{43E8275B-683A-45EB-9BB5-34F85F7E30AF}</Project>
-      <Name>AsyncSvc_sample</Name>
-    </ProjectReference>
-  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
