+++ root/programs/CS/Samples/AsyncSvc_sample/TestAsyncSvc_Sample/app.config
@@ -3,7 +3,7 @@
   <appSettings>
     <!-- Log4Netのコンフィグファイルへのパス -->
     <!-- リソース ファイルの場合 -->
-    <add key="FxLog4NetConfFile" value="C:\root\files\resource\Log\SampleLogConf.xml" />
+    <add key="FxLog4NetConfFile" value="C:\root\files\resource\AsyncProcessingService\Log\SampleLogConf.xml" />
     <!-- 埋め込まれたリソースの場合 -->
     <!-- Ｄ層のパフォーマンスログ出力機能のon・off -->
     <add key="FxSqlTraceLog" value="on" />
@@ -17,19 +17,11 @@
     <!-- 共通部品の使用するパラメータ - end -->
     <!-- アプリケーションの使用するパラメータ - start -->
     <!-- SQLファイルファイル（フォルダ）へのパス -->
-    <add key="SqlTextFilePath" value="C:\OpenTouryo\root\files\resource\Sql\sqlserver\AsyncProcessingService"/>
+    <add key="SqlTextFilePath" value="C:\root\files\resource\AsyncProcessingService\Sql\sqlserver"/>
     <!-- アプリケーションの使用するパラメータ - end -->
   </appSettings>
   <connectionStrings>
     <!-- SQL Server / SQL Client用 -->
     <add name="ConnectionString_SQL" connectionString="Data Source=localhost\sqlexpress;Integrated Security=SSPI;Initial Catalog=AsyncProcessingServiceDB;"/>
   </connectionStrings>
-    <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-10.0.0.0" newVersion="10.0.0.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
 </configuration>
