+++ root/programs/VB/Samples/AsyncSvc_sample/TestAsyncSvc_Sample/app.config
@@ -1,7 +1,6 @@
 <?xml version="1.0"?>
 <configuration>
   <appSettings>
-
     <!-- Log4Netのコンフィグファイルへのパス -->
     <!-- リソース ファイルの場合 -->
     <add key="FxLog4NetConfFile" value="C:\root\files\resource\Log\SampleLogConf.xml" />
@@ -15,20 +14,14 @@
     <add key="FxSqlEncoding" value="utf-8"/>
     <!-- Ｄ層のコマンド タイムアウト値を指定（秒） -->
     <add key="FxSqlCommandTimeout" value="30"/>
-
     <!-- 共通部品の使用するパラメータ - end -->
-
     <!-- アプリケーションの使用するパラメータ - start -->
-
     <!-- SQLファイルファイル（フォルダ）へのパス -->
     <add key="SqlTextFilePath" value="C:\OpenTouryo\root\files\resource\Sql\sqlserver\AsyncProcessingService"/>
-
     <!-- アプリケーションの使用するパラメータ - end -->
-
   </appSettings>
   <connectionStrings>
     <!-- SQL Server / SQL Client用 -->
     <add name="ConnectionString_SQL" connectionString="Data Source=localhost\sqlexpress;Initial Catalog=AsyncProcessingServiceDB;Integrated Security=True"/>
-
   </connectionStrings>
 </configuration>
